Move ModelNotReady to errors.py
