+++ python/kserve/kserve/errors.py
@@ -58,6 +58,17 @@ class ModelNotFound(Exception):
         return self.reason
 
 
+class ModelNotReady(RuntimeError):
+    def __init__(self, model_name: str, detail: str = None):
+        self.model_name = model_name
+        self.error_msg = f"Model with name {self.model_name} is not ready."
+        if detail:
+            self.error_msg = self.error_msg + " " + detail
+
+    def __str__(self):
+        return self.error_msg
+
+
 async def exception_handler(_, exc):
     return JSONResponse(status_code=HTTPStatus.INTERNAL_SERVER_ERROR, content={"detail": str(exc)})
