+++ python/kserve/kserve/handlers/model_repository_extension.py
@@ -2,19 +2,8 @@ import inspect
 import sys
 from typing import Dict, List
 
-from kserve.errors import ModelNotFound
+from kserve.errors import ModelNotFound, ModelNotReady
 from kserve import ModelRepository
-
-
-class ModelNotReady(RuntimeError):
-    def __init__(self, model_name: str, detail: str = None):
-        self.model_name = model_name
-        self.error_msg = f"Model with name {self.model_name} is not ready."
-        if detail:
-            self.error_msg = self.error_msg + " " + detail
-
-    def __str__(self):
-        return self.error_msg
 
 
 class ModelRepositoryExtension:
