+++ python/kserve/test/test_model_repository_extension.py
@@ -1,6 +1,6 @@
 import pytest
-from kserve.errors import ModelNotFound
-from kserve.handlers.model_repository_extension import ModelRepositoryExtension, ModelNotReady
+from kserve.errors import ModelNotFound, ModelNotReady
+from kserve.handlers.model_repository_extension import ModelRepositoryExtension
 from kserve.model_repository import ModelRepository
 from test.test_server import DummyModel, DummyModelRepository
